% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_tribble.R
\name{df_to_tribble}
\alias{df_to_tribble}
\title{Convert a reactive data frame in Shiny to a tribble for editor}
\usage{
df_to_tribble(.data, .name = NULL)
}
\arguments{
\item{.data}{A data frame to convert}

\item{.name}{Optional name for tribble output}
}
\value{
Text sting that can be read as a tribble in R
}
\description{
Convert a reactive data frame in Shiny to a tribble for editor
Mostly for use inside Shiny RStudio Addins
}
