% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ-data.R
\docType{data}
\name{GVAIndustry}
\alias{GVAIndustry}
\title{Gross Value Added & Employment by Industry}
\format{An object of class \code{"data.frame"}.}
\usage{
data(GVAIndustry)
}
\description{
Gross Value Added & Employment by Industry
}
\keyword{datasets}
