% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspivot.R
\name{rspivot}
\alias{rspivot}
\title{View data frames as Shiny pivot tables}
\usage{
rspivot(df = .Last.value, valueName = "value")
}
\arguments{
\item{df}{A data frame flat file to be converted to pivot format. Data should be in "long" format, with a single column of values named \code{value}.
The function defaults to showing the most recent object made in R.}

\item{valueName}{Name of series in \code{df} containing data values. Defaults to "value".
Can also accept an array of strings containing the names of multiple value columns in the data frame.}
}
\value{
An RStudio dialog box will pop-up with a Shiny pivot table of the data.
}
\description{
View data frames as Shiny pivot tables. This is an alternative to \code{View()} to view data frames as summarized data.
}
\examples{
rspivot(GVAIndustry)


GVAIndustry2 <- GVAIndustry \%>\%
     spread(Econ, value)

rspivot(GVAIndustry2, valueName = c("Employment", "GDP"))
}
