% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvaindustry-data.R
\docType{data}
\name{GVAIndustry}
\alias{GVAIndustry}
\title{Eurostat Gross Value Added & Other Series by Industry}
\format{A data frame with 136537 rows and 5 descriptive variables:
\describe{
\item{Country}{European country, 34 total}
\item{Price}{price metric, Real (base year 2010) or Nominal, should not be summed}
\item{Series}{national accounts series, 9 total, use caution when summing}
\item{Industy}{industry ISIC Rev.3.1, 21 top level industries}
\item{Year}{year, 1996-2016, earlier & recent years may not be complete}
}}
\source{
\url{http://ec.europa.eu/eurostat/data/database}
}
\usage{
GVAIndustry
}
\description{
A dataset containing gross value added and other national statistics by industry for Eurostat countries.
}
\keyword{datasets}
