% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspivot.R
\name{rspivot}
\alias{rspivot}
\title{View data frames as Shiny pivot tables}
\usage{
rspivot(df = .Last.value, valueName = "value", initCols = "",
  initRows = "", initNest = "", initFilters = list(),
  initMetric = list(metric = "Values", series = ""))
}
\arguments{
\item{df}{A data frame flat file to be converted to pivot format. Data should be in "long" format, with a single column of values named \code{value}.
The function defaults to showing the most recent object made in R.}

\item{valueName}{Name of series in \code{df} containing data values. Defaults to "value".
Can also accept an array of strings containing the names of multiple value columns in the data frame.}

\item{initCols}{Specify the series to be displayed as columns. If blank, defaults to the right-most series in the data frame.}

\item{initRows}{Specify the series to be displayed as rows. If blank, defaults to the 2nd right-most series in the data frame.}

\item{initNest}{Specify the series to be displayed as nested rows. If blank, no nested rows are displayed.}

\item{initFilters}{Optional list of initial filter selections. Leave a series blank or use "Show All" to select all. Pass series names to \code{make.names()} to ensure correct use.
Alternatively, leave this blank on the intiial run, and use the Save Function feature after manually selecting filters.}

\item{initMetric}{Optional list of the initial data metrics to display.}
}
\description{
View data frames as Shiny pivot tables. This is an alternative to \code{View()} to view data frames as summarized data.
Additional options include sparklines and charts and multiple data metrics.
}
\examples{
\dontrun{
rspivot(GVAIndustry)

GVAIndustry2 <- GVAIndustry \%>\%
     spread(Econ, value)
rspivot(GVAIndustry2, valueName = c("Employment", "GDP"))
}
}
